backend/gateway_service/pyproject.toml
[project]
name = "gateway-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "python-jose[cryptography]"
]
